Implement live streaming for execute resources
This brings live streaming of execute resource output to the
output formatters. It also adds a mechanism for checking to
see if an output formatter is in use through the event dispatch
system.

It adds a new configuration option, "always_stream_execute", which
does what it says on the tin.